Add files via upload
I added a larger buffer (2*forecast_horizon) between the end of the train data and the beginning of the test data. This makes sure that the test data is completely isolated from the train data. This change also made it necessary to put the forecast_horizon parameter into the evaluate_model_1  function.