Add files via upload
I put the use functions in there own files and changed their names. I also took away a +1 in pred_start_day for both the solar model function and the use model function. The length of time between the train and test data should now be two weeks, as desired.